+++ bench/toa.js
@@ -10,9 +10,7 @@ var n = parseInt(process.env.MW || '1', 10)
 process.stdout.write('  toa, ' + n + ' middleware:')
 
 while (n--) {
-  app.use(function * () {
-    yield (done) => setImmediate(done) // fake task
-  })
+  app.use(function * () {})
 }
 
 var body = new Buffer('Hello World')