+++ test/context.js
@@ -4,7 +4,6 @@
 // **License:** MIT
 
 var Stream = require('stream')
-var thunks = require('thunks')
 var toa = require('..')
 
 exports = module.exports = function (req, res) {
@@ -28,7 +27,8 @@ exports = module.exports = function (req, res) {
   res.removeHeader = function (k, v) {
     delete res._headers[k.toLowerCase()]
   }
-  return toa.createContext(toa(), req, res, thunks())
+  var app = toa()
+  return new app.Context(req, res)
 }
 
 exports.context = function (req, res) {